Create route group for main routes Update css for individual blog page Add semantic metadata for routes
